+++ HISTORY.md
@@ -13,6 +13,7 @@
 - Adding `mypy` checks to CI.
 - Adding `--dry-run` and `--show-diff` arguments to `cubi-sak snappy pull-sheet`.
 - Adding `cubi-sak snake check` command.
+- Adding `cubi-sak isa-tab-validate` command.
 
 ## v0.1.0